+++ app/backend/enigma/src/main/java/com/wia/enigma/configuration/security/JwtRequestFilter.java
@@ -1,12 +1,9 @@
 package com.wia.enigma.configuration.security;
 
-import com.wia.enigma.core.service.EnigmaJwtService;
-import com.wia.enigma.dal.enums.AudienceType;
+import com.wia.enigma.core.service.JwtService.EnigmaJwtService;
 import com.wia.enigma.dal.enums.ExceptionCodes;
 import com.wia.enigma.exceptions.custom.EnigmaUnauthorizedException;
-import com.wia.enigma.utilities.AuthUtils;
 import com.wia.enigma.utilities.JwtUtils;
-import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,7 +15,6 @@ import org.springframework.http.HttpMethod;
 import org.springframework.lang.NonNull;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
