+++ app/backend/enigma/src/main/java/com/wia/enigma/core/controller/api/v1/EnigmaUserController.java
@@ -1,9 +1,7 @@
 package com.wia.enigma.core.controller.api.v1;
 
 import com.wia.enigma.configuration.security.EnigmaAuthenticationToken;
-import com.wia.enigma.core.service.EnigmaUserServiceImpl;
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
