+++ app/backend/enigma/src/main/java/com/wia/enigma/core/service/JwtService/EnigmaJwtService.java
@@ -1,4 +1,4 @@
-package com.wia.enigma.core.service;
+package com.wia.enigma.core.service.JwtService;
 
 import com.wia.enigma.configuration.security.EnigmaAuthenticationToken;
 import com.wia.enigma.configuration.security.EnigmaUserDetails;
