+++ app/backend/enigma/src/main/java/com/wia/enigma/core/service/VerificationTokenService/VerificationTokenServiceImpl.java
@@ -1,6 +1,7 @@
-package com.wia.enigma.core.service;
+package com.wia.enigma.core.service.VerificationTokenService;
 
 
+import com.wia.enigma.core.service.VerificationTokenService.VerificationTokenService;
 import com.wia.enigma.dal.entity.VerificationToken;
 import com.wia.enigma.dal.enums.ExceptionCodes;
 import com.wia.enigma.dal.repository.VerificationTokenRepository;
